feat: merge finances levels
